Job deployed to nats connects to nats on own IP
During upgrades, if previous deployment does not have nats-tls or the
metrics-discovery-registrar job at all, deployment can fail because
nats.service.cf.internal does not return any healthy instances to
connect that also have nats-tls running.

The first nats intance to deploy that would have that job running is not
healthy until the metrics-discovery-registrar is running. But
metrics-discovery-registrar has no nats-tls job it can resolve with
bosh-dns because the only instance that has nats-tls running is the
same, first nats instance. But that instance isn't healthy yet because
the metrics-discovery-registrar job isn't running yet... you get the
idea.

One way around this is to change the bosh-dns-aliases query from "*" to
"q-s4". But that means anyone locating nats will get unhealthy instances
in the mix. That may be acceptable given how nats clients retry
connections. But this change is more localized and addresses the root
cause more directly.

With this change, when nats_instance_group is configured, the
metrics-discovery-registrar jobs that get deployed to the instance group
of that name will connect to the nats job running on the same VM by the
bosh-dns address of that particular instance, rather than a bosh-dns
query that matches a random healthy instance. The new property defaults
to an empty string, so if it is not set, the current behavior is
preserved.