Fix failing file writer test
After a ginkgo/gomega bump, this test started failing. Appeared to have
been due to a change where when that test was run in parallel, the
shared scrapeCfgs variable wasn't being reset across processes, so tests
that expected them to be blank were failing.

Add a BeforeEach node to rest scrapeCfgs to an empty slice.